f2fs: Switch long to int type for f2fs_invalidate_page
../../../../../../kernel/xiaomi/libra/fs/f2fs/checkpoint.c:397:20: warning: incompatible function pointer types initializing 'void (*)(struct page *, unsigned int, unsigned int)' with an expression of type 'void (struct page *, unsigned long, unsigned int)' [-Wincompatible-function-pointer-types]
        .invalidatepage = f2fs_invalidate_page,
                          ^~~~~~~~~~~~~~~~~~~~
../../../../../../kernel/xiaomi/libra/fs/f2fs/data.c:2309:20: warning: incompatible function pointer types initializing 'void (*)(struct page *, unsigned int, unsigned int)' with an expression of type 'void (struct page *, unsigned long, unsigned int)' [-Wincompatible-function-pointer-types]
        .invalidatepage = f2fs_invalidate_page,
                          ^~~~~~~~~~~~~~~~~~~~
2 warnings generated.

+++ fs/f2fs/f2fs.h
@@ -181,9 +181,9 @@ static inline void inode_unlock(struct inode *inode)
  *
  * Please refer to the comment for waitqueue_active.
  */
-static inline bool wq_has_sleeper(wait_queue_head_t *wq) 
+static inline bool wq_has_sleeper(wait_queue_head_t *wq)
 {
-	/*   
+	/*
 	 * We need to be sure we are in sync with the
 	 * add_wait_queue modifications to the wait queue.
 	 *
@@ -215,8 +215,8 @@ static inline struct timespec current_time(struct inode *inode)
 
 	if (unlikely(!inode->i_sb)) {
 		WARN(1, "current_time() called with uninitialized super_block in the inode");
-		return now; 
-	}    
+		return now;
+	}
 
 	return timespec_trunc(now, inode->i_sb->s_time_gran);
 }
@@ -2845,7 +2845,7 @@ int f2fs_map_blocks(struct inode *inode, struct f2fs_map_blocks *map,
 int f2fs_fiemap(struct inode *inode, struct fiemap_extent_info *fieinfo,
 			u64 start, u64 len);
 void f2fs_set_page_dirty_nobuffers(struct page *page);
-void f2fs_invalidate_page(struct page *page, unsigned long offset, unsigned int length);
+void f2fs_invalidate_page(struct page *page, unsigned int offset, unsigned int length);
 
 int __f2fs_write_data_pages(struct address_space *mapping,
 						struct writeback_control *wbc,
